feat: complete Phase 5 - restructure entry points and consolidate CLI architecture
  Move memory_system utilities to heimdall/cognitive_system/, update entry points
  to heimdall/heimdall-mcp, fix all imports, and update documentation. Achieves
  unified CLI structure with clean operations layer separation.
