+++ heimdall/cli_commands/health_commands.py
@@ -14,8 +14,12 @@ from cognitive_memory.main import (
     initialize_system,
     initialize_with_config,
 )
+from heimdall.cognitive_system.health_checker import (
+    HealthChecker,
+    HealthCheckResults,
+    HealthResult,
+)
 from heimdall.interactive_shell import InteractiveShell
-from memory_system.health_checker import HealthChecker, HealthCheckResults, HealthResult
 
 console = Console()
