+++ heimdall/cognitive_system/__init__.py
@@ -1,0 +1,1 @@
+"""Cognitive system utilities and services."""
