--- memory_system/__init__.py
@@ -1,8 +1,0 @@
-"""
-Unified memory system CLI package.
-
-This package provides the main entry point for the cognitive memory system
-with embedded service management, health checking, and multiple interface support.
-"""
-
-__version__ = "0.1.0"
