+++ scripts/release.py
@@ -53,7 +53,7 @@ def validate_package() -> bool:
     # Check if package can be imported
     try:
         run_command(
-            "python -c 'import cognitive_memory; import interfaces; import memory_system'"
+            "python -c 'import cognitive_memory; import interfaces; import heimdall'"
         )
     except SystemExit:
         console.print("[red]Error: Package modules cannot be imported[/red]")
